requirements.txt
beautifulsoup4==4.10.0
community==1.0.0b1
dhlab>=2.0.0
#ipython==8.1.1
matplotlib==3.4.3
#networkx==2.6.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.1
#python_louvain==0.16
requests==2.26.0
seaborn==0.11.2
streamlit>=1.10.0
#wordcloud==1.8.1
